refactor: clean up discojs import and tighten numeric input validation
Use bare 'discojs' import instead of fragile deep path, simplify
buildPagination with conditional spreads, add .int() to folderId/page/perPage.

+++ src/server/trpc/routers/discogs.ts
@@ -97,9 +97,9 @@ export const discogsRouter = router({
     .input(
       optionalAuthInput.extend({
         username: z.string(),
-        folderId: z.number().min(0).optional().default(0),
-        page: z.number().min(1).optional().default(1),
-        perPage: z.number().min(1).max(100).optional().default(50),
+        folderId: z.number().int().min(0).optional().default(0),
+        page: z.number().int().min(1).optional().default(1),
+        perPage: z.number().int().min(1).max(100).optional().default(50),
         sort: z.enum(USER_SORT_VALUES).optional(),
         sortOrder: z.enum(SORT_ORDER_VALUES).optional()
       })
